+++ apps/dashboard/app/controllers/jobs/projects_controller.rb
@@ -1,0 +1,9 @@
+class Jobs::ProjectsController < ApplicationController
+  def show
+    # files in current project
+  end
+
+  def index
+    @projects = Jobs::Project.all
+  end
+end