+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   in [1611](https://github.com/OSC/ondemand/pull/1611).
 - Changed the file editor navbar in [1582](https://github.com/OSC/ondemand/pull/1582).
 - CI now uses the `ood_packaging` gem to build artifacts in [1661](https://github.com/OSC/ondemand/pull/1661).
+- Yaml files now preview in [1758](https://github.com/OSC/ondemand/pull/1758).
 
 ### Added